dtsi: Set default flashlight current to maximum supported
This is needed to fix low flashlight brightness on custom ROMs because they do not have a custom flashlight brightness adjuster like ONEUI.

TEST: Works for A10
Remarks: On oneui, this value is overridden by the torch brightness slider interface and works fine. On custom roms, this seems to fix the low flashlight issue.
Conclusion: Success!
Change-Id: I989f53dc201714fb04f24947996f47d4ad3f23fa
Signed-off-by: Chatur27 <jasonbright2709@gmail.com>